CORE-6662: Eliminate duplicated code in commands
This is a refactoring change, functionally this is a NOP.
