+++ tools/plugins/initial-rbac/src/main/kotlin/net/corda/cli/plugin/initialRbac/commands/RoleCreationUtils.kt
@@ -1,0 +1,80 @@
+package net.corda.cli.plugin.initialRbac.commands
+
+import net.corda.cli.plugins.common.HttpRpcClientUtils
+import net.corda.cli.plugins.common.HttpRpcClientUtils.createHttpRpcClient
+import net.corda.cli.plugins.common.HttpRpcCommand
+import net.corda.libs.permissions.endpoints.v1.permission.PermissionEndpoint
+import net.corda.libs.permissions.endpoints.v1.permission.types.CreatePermissionType
+import net.corda.libs.permissions.endpoints.v1.permission.types.PermissionType
+import net.corda.libs.permissions.endpoints.v1.role.RoleEndpoint
+import net.corda.libs.permissions.endpoints.v1.role.types.CreateRoleType
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+import java.time.Duration
+import java.time.temporal.ChronoUnit
+
+internal object RoleCreationUtils {
+
+    /**
+     * Checks if role already exists and then does nothing, else:
+     * - creates permissions;
+     * - creates role;
+     * - assigns permissions to the role.
+     */
+    fun HttpRpcCommand.checkOrCreateRole(roleName: String, permissionsToCreate: Map<String, String>): Int {
+
+        val logger: Logger = LoggerFactory.getLogger(this::class.java)
+        val sysOut: Logger = LoggerFactory.getLogger("SystemOut")
+        val errOut: Logger = LoggerFactory.getLogger("SystemErr")
+
+        logger.info("Running ${this.javaClass.simpleName}")
+
+        createHttpRpcClient(RoleEndpoint::class).use { roleEndpointClient ->
+            val waitDuration = Duration.of(waitDurationSeconds.toLong(), ChronoUnit.SECONDS)
+            val roleEndpoint = HttpRpcClientUtils.executeWithRetry(waitDuration, "Start of role HTTP endpoint") {
+                roleEndpointClient.start().proxy
+            }
+            val allRoles = HttpRpcClientUtils.executeWithRetry(waitDuration, "Obtain list of available roles") {
+                roleEndpoint.getRoles()
+            }
+            if (allRoles.any { it.roleName == roleName }) {
+                errOut.error("$roleName already exists - nothing to do.")
+                return 5
+            }
+
+            val permissionIds = createHttpRpcClient(PermissionEndpoint::class).use { permissionEndpointClient ->
+                val permissionEndpoint = permissionEndpointClient.start().proxy
+                permissionsToCreate.toSortedMap().map { entry ->
+                    HttpRpcClientUtils.executeWithRetry(waitDuration, "Creating permission: ${entry.key}") {
+                        permissionEndpoint.createPermission(
+                            CreatePermissionType(
+                                PermissionType.ALLOW,
+                                entry.value,
+                                null,
+                                null
+                            )
+                        )
+                    }.responseBody.id.also {
+                        logger.info("Created permission: ${entry.key} with id: $it")
+                    }
+                }
+            }
+
+            val roleId = HttpRpcClientUtils.executeWithRetry(waitDuration, "Creating role: $roleName") {
+                roleEndpoint.createRole(CreateRoleType(roleName, null)).responseBody.id
+            }
+            permissionIds.forEach { permId ->
+                HttpRpcClientUtils.executeWithRetry(
+                    waitDuration,
+                    "Adding permission: $permId",
+                    onAlreadyExists = HttpRpcClientUtils::ignore
+                ) {
+                    roleEndpoint.addPermission(roleId, permId)
+                }
+            }
+            sysOut.info("Successfully created $roleName with id: $roleId and assigned permissions")
+        }
+
+        return 0
+    }
+}
